Added "Bigger icons in calendar view"
Added a Markdown file for the "Bigger icons in calendar view" CSS Tweak - essentially just an export from the CSS Tweaks page in Notion. Conveniently, I'd been linking to image files on Discord's CDN, so they needn't be hosted in this repository. I also added a hyperlink to the README file for the Tweak.

@dragonwocky mainly just wanted to see if you are happy with this format. thank you!